Fixing docs previews generating
