+++ lib/tasks/docs.rake
@@ -378,14 +378,13 @@ namespace :docs do
       path.dirname.mkdir unless path.dirname.exist?
 
       File.open(path, "w") do |f|
-        f.puts("module Primer")
-        f.puts("  module Docs")
-        f.puts("    class #{short_name}Preview < ViewComponent::Preview")
+        f.puts("module Docs")
+        f.puts("  class #{short_name}Preview < ViewComponent::Preview")
 
         yard_example_tags.each_with_index do |tag, index|
           name, _, code = parse_example_tag(tag)
           method_name = name.split("|").first.downcase.parameterize.underscore
-          f.puts("      def #{method_name}; end")
+          f.puts("    def #{method_name}; end")
           f.puts unless index == yard_example_tags.size - 1
           path = Pathname.new("lookbook/test/components/previews/docs/#{short_name.underscore}_preview/#{method_name}.html.erb")
           path.dirname.mkdir unless path.dirname.exist?
@@ -394,7 +393,6 @@ namespace :docs do
           end
         end
 
-        f.puts("    end")
         f.puts("  end")
         f.puts("end")
       end
