+++ lookbook/test/components/previews/counter_component_preview.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+# no doc
+class CounterComponentPreview < ViewComponent::Preview
+  # @label Rounded
+  #
+  # @param scheme [Symbol] select [[Default, default], [Primary, primary], [Secondary, secondary]]]
+  def default(scheme: :default)
+    render(Primer::CounterComponent.new(count: 2, scheme: scheme))
+  end
+end
