--- lookbook/test/components/previews/primer/counter_component_preview.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-module Primer
-  # no doc
-  class CounterComponentPreview < ViewComponent::Preview
-    # @label Rounded
-    #
-    # @param scheme [Symbol] select [[Default, default], [Primary, primary], [Secondary, secondary]]]
-    def default(scheme: :default)
-      render(Primer::CounterComponent.new(count: 2, scheme: scheme))
-    end
-  end
-end
