--- lookbook/test/components/previews/primer/tooltip_preview.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-module Primer
-  # no doc
-  class TooltipPreview < ViewComponent::Preview
-    def description; end
-
-    def label; end
-  end
-end
