+++ lookbook/test/components/previews/time_ago_component_preview.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+# no doc
+class TimeAgoComponentPreview < ViewComponent::Preview
+  def default
+    render(Primer::TimeAgoComponent.new(time: Time.zone.now))
+  end
+
+  def micro
+    render(Primer::TimeAgoComponent.new(time: Time.zone.now, micro: true))
+  end
+end
