+++ lookbook/test/components/previews/tooltip_preview.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+# no doc
+class TooltipPreview < ViewComponent::Preview
+  def description; end
+
+  def label; end
+end
